+++ lib/Api/SmartTransactionsApi.php
@@ -61,6 +61,89 @@ class SmartTransactionsApi
     }
 
     /**
+     * Operation addTransaction
+     *
+     * Create Smart Transaction
+     *
+     * @param \Secuconnect\Client\Model\SmartTransactionsDTO $body Transaction details
+     * @throws ApiException on non-2xx response
+     * @return \Secuconnect\Client\Model\SmartTransactionsProductModel
+     */
+    public function addTransaction($body)
+    {
+        list($response) = $this->addTransactionWithHttpInfo($body);
+        return $response;
+    }
+
+    /**
+     * Operation addTransactionWithHttpInfo
+     *
+     * Create Smart Transaction
+     *
+     * @param null|\Secuconnect\Client\Model\SmartTransactionsDTO $body Transaction details
+     * @throws ApiException on non-2xx response
+     * @return array<int, mixed> of \Secuconnect\Client\Model\SmartTransactionsProductModel, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function addTransactionWithHttpInfo($body)
+    {
+        // parse inputs
+        $resourcePath = "/Smart/Transactions";
+        $httpBody = [];
+        $queryParams = [];
+        $headerParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType(['application/json']);
+
+
+        // body params
+        $httpBody = $body;
+
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'POST',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    '\Secuconnect\Client\Model\SmartTransactionsProductModel',
+                    '/Smart/Transactions'
+                );
+
+                return [$this->apiClient->getSerializer()->deserialize($response, '\Secuconnect\Client\Model\SmartTransactionsProductModel', $httpHeader), $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                    case 200:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\SmartTransactionsProductModel', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                    case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
+
+    /**
      * Operation getAll
      *
      * Find Smart Transactions
@@ -374,6 +457,9 @@ class SmartTransactionsApi
             $resourcePath
         );
 
+        // body params
+        $httpBody = $body;
+
         for ($retries = 0; ; $retries++) {
 
             // this endpoint requires OAuth (access token)
@@ -467,6 +553,9 @@ class SmartTransactionsApi
             $this->apiClient->getSerializer()->toPathValue($smart_transaction_id),
             $resourcePath
         );
+
+        // body params
+        $httpBody = $body;
 
         for ($retries = 0; ; $retries++) {